Move DUAL_STICK_MODE const
