+++ www/src/Data/Addons.ts
@@ -1,0 +1,5 @@
+export const DUAL_STICK_MODES = [
+	{ label: 'D-Pad', value: 0 },
+	{ label: 'Left Analog', value: 1 },
+	{ label: 'Right Analog', value: 2 },
+];
